Add configuration management endpoints and models
- Implement POST endpoint to save application configuration.
- Create ConfigModel for application configuration.
- Introduce ModelInfo structure for model information.
