+++ backend/app/api/v1/endpoints/config.py
@@ -7,6 +7,8 @@ from fastapi import APIRouter, HTTPException
 
 from app.schemas.config import ConfigResponse
 from app.core.config import settings
+from app.models.config import ConfigModel
+import json
 
 logger = logging.getLogger("app.api.config")
 router = APIRouter(prefix="/config", tags=["configuration"])
@@ -29,3 +31,16 @@ async def get_config() -> ConfigResponse:
         ollama_url=settings.ollama_url,
         use_langchain_router=settings.use_langchain_router
     )
+
+
+# POST endpoint to save configuration
+@router.post("/", response_model=dict)
+async def set_config(cfg: ConfigModel):
+    """
+    Save application configuration to a file.
+    """
+    logger.info(f"POST /config called with: {cfg}")
+    with open("config.json", "w") as f:
+        json.dump(cfg.model_dump(), f, indent=2)
+    logger.info("Config saved.")
+    return {"status": "ok"}
